refactor: updating and refactoring api
