+++ src/server/requests/base_test.http
@@ -1,0 +1,3 @@
+### GET base
+http://localhost:8080/api
+
